Add files via upload
This project is not an original ides. It was copied from another that I saw on GitHub and wanted to see if I can recreate it. It is a layout for a small office. This is the original and I would make more changes to it and post it on here as the time goes