Add files via upload
In this version, the IP addresses are assigned by a DHCP server allocated to each department, unlike in the previous version, where the devices were unable to communicate with each other; this version has communication between devices in different departments. I know this is not ideal due to things like maintenance and the like, so I will try to make a better version where only one server makes the IPs for the whole organisation later down the line.